k8s: Kubernetes resources for benchmark binary
This commit adds Kubernetes resources for running the benchmark binary
as a Pod.

Adds the following resources manifests:

ServiceAccount to run the benchmark binary in privileged
mode, we use a different service account.

RoleBinding for attaching the `zz-privileged` psp to the above service
account

ClusterRole and ClusterRoleBinding to be able to create calico related
resources from inside the benchmark pod.

Pod manifest for benchmark.

Signed-off-by: Imran Pochi <imran@kinvolk.io>